1.) README & Lizenz angepasst 2.) Doku hoffentlich aktuell 3.) Warnungen eliminiert

+++ src/JSON_parser.h
@@ -25,9 +25,9 @@
 
 /* Determine the integer type use to parse non-floating point numbers */
 #if __STDC_VERSION__ >= 199901L || HAVE_LONG_LONG == 1
-typedef long long JSON_int_t;
-#define JSON_PARSER_INTEGER_SSCANF_TOKEN "%lld"
-#define JSON_PARSER_INTEGER_SPRINTF_TOKEN "%lld"
+typedef long JSON_int_t;
+#define JSON_PARSER_INTEGER_SSCANF_TOKEN "%ld"
+#define JSON_PARSER_INTEGER_SPRINTF_TOKEN "%ld"
 #else 
 typedef long JSON_int_t;
 #define JSON_PARSER_INTEGER_SSCANF_TOKEN "%ld"
